# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Swagger API.",
        "title": "Gin Swagger Demo",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8888",
    "paths": {
        "/huangcoolyserve/empolyee/getEmployee": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "取得Employee人員",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User Id",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/huangcoolyserver/empolyee/addEmployee": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Add a new user to the store",
                "parameters": [
                    {
                        "description": "User Data",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/paramDto.ParamEmployeeDto"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/huangcoolyserver/empolyee/updateEmployee": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Add a new user to the store",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Employee name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "沒填寫預設則為0",
                        "name": "age",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/huangcoolyserver/user/checkUserLoginStatus": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "驗證使用者的登陸狀態",
                "parameters": [
                    {
                        "description": "使用者資訊",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/paramDto.ParamUserIdDto"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/huangcoolyserver/user/getRoleAccessSystem/{userId}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "取得使用者可以使用的Navbar",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Id",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/huangcoolyserver/user/getUser": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "取得User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User Id",
                        "name": "userId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/huangcoolyserver/user/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "使用者進行登入",
                "parameters": [
                    {
                        "description": "使用者資訊",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/paramDto.ParamUserLoginDto"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "paramDto.ParamEmployeeDto": {
            "description": "User account information with user id and username",
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "paramDto.ParamUserIdDto": {
            "description": "User account information with user id and username",
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string"
                },
                "userPassword": {
                    "type": "string"
                }
            }
        },
        "paramDto.ParamUserLoginDto": {
            "description": "使用者登入資訊 with user id and username",
            "type": "object",
            "properties": {
                "userPassword": {
                    "type": "string"
                },
                "userWorkId": {
                    "type": "string"
                }
            }
        }
    }
}